+++ src/Idris/Termination.hs
@@ -1,6 +1,6 @@
 {-|
-Module      : Idris.Coverage
-Description : The coverage and totality checkers for Idris are in this module.
+Module      : Idris.Termination
+Description : The termination checker for Idris
 Copyright   :
 License     : BSD3
 Maintainer  : The Idris Community.